Move AuxmapsModel and AuxmapsMixin to model_auxmaps.py
