+++ hydromt/models/__init__.py
@@ -6,6 +6,7 @@ from .model_api import *
 from .model_grid import *
 from .model_lumped import *
 from .model_network import *
+from .model_auxmaps import *
 from . import model_plugins
 
 from .. import _has_xugrid
