+++ tests/test_model.py
@@ -6,8 +6,8 @@ import pytest
 import xarray as xr
 import numpy as np
 import geopandas as gpd
-from hydromt.models.model_api import _check_data, AuxmapsModel
-from hydromt.models import Model, GridModel, LumpedModel
+from hydromt.models.model_api import _check_data
+from hydromt.models import Model, GridModel, LumpedModel, AuxmapsModel
 from hydromt import _has_xugrid
 
 # from .conftest import TestAuxModel
